Fix ConnectionTracker: AtomicLong.get().toInt() for connectionCount
AtomicLong.get() returns Long but ConnectionEntry.connectionCount is
Int, causing a compile error. Add .toInt().

+++ app/src/main/kotlin/com/multivpn/app/network/ConnectionTracker.kt
@@ -73,7 +73,7 @@ class ConnectionTracker(context: Context) {
      */
     fun snapshot(): List<ConnectionEntry> {
         val entries = domains.map { (domain, acc) ->
-            ConnectionEntry(domain, acc.tx.get(), acc.rx.get(), acc.count.get())
+            ConnectionEntry(domain, acc.tx.get(), acc.rx.get(), acc.count.get().toInt())
         }.sortedByDescending { it.uploadedBytes + it.downloadedBytes }
 
         if (!tracking) return entries
